fixup! Big refactor...
