+++ tests/myproject/myproject/flows/batch_flow.py
@@ -1,17 +1,14 @@
+from pathlib import Path
+
 import daps_utils  # This works on batch because preinstall-batch_flow.sh installs it
 from metaflow import batch, FlowSpec, step
 
 
-class BatchFlow(FlowSpec, daps_utils.DapsFlowMixin):
+class MetaflowExtensionsPreinstallFlow(FlowSpec, daps_utils.DapsFlowMixin):
     """Dummy batch flow for showing that bootstrap_command is working."""
 
     @step
     def start(self):
-        """Start the flow."""
-        self.next(self.a)
-
-    @step
-    def a(self):
         """Prepare chunks for batching."""
         self.items = [1]
         self.next(self.b, foreach="items")
@@ -19,13 +16,11 @@ class BatchFlow(FlowSpec, daps_utils.DapsFlowMixin):
     @batch()
     @step
     def b(self):
-        """Show that daps_utils is installed via preinstall."""
+        """Show that daps_utils is installed via. preinstall."""
         self.something = self.input
-        # Show daps_utils is installed via preinstall
         print("non-conda batch task has installed", daps_utils.__version__)
-        # Show preinstall has run (it adds a file called 'special-batch-file')
-        from pathlib import Path
 
+        # Show preinstall has run (it adds a file called 'special-batch-file')
         assert Path("special-batch-file").exists()
         self.next(self.c)
 
@@ -41,4 +36,4 @@ class BatchFlow(FlowSpec, daps_utils.DapsFlowMixin):
 
 
 if __name__ == "__main__":
-    BatchFlow()
+    MetaflowExtensionsPreinstallFlow()
